Revert to the plain Z shell prompt
Under heavy load, the prompt is noticeably slow. Also, it can display
a lot of information, but I actually don't really care about it.
Unless you type commands all the time, the prompt will be outdated
(taking the example of a git repo). But when you are typing commands
all the time, you know the state of the repo… Superfluous to me.